+++ packages/deploy/index.ts
@@ -1,0 +1,2 @@
+export { NocturneDeployer, NocturneDeployOpts } from "./src/deploy";
+export { checkNocturneDeploymentConfig } from "./src/checks";