+++ packages/e2e-tests/test/FullEndToEnd.ts
@@ -23,8 +23,8 @@ import {
   calculateOperationDigest,
   AssetType,
 } from "@nocturne-xyz/sdk";
-import { setupNocturne } from "../deploy/deployNocturne";
-import { depositFunds, sleep, getSubtreeUpdateProver } from "./utils";
+import { setupNocturne } from "../utils/deploy";
+import { depositFunds, sleep, getSubtreeUpdateProver } from "../utils/test";
 import { OperationProcessedEvent } from "@nocturne-xyz/contracts/dist/src/Wallet";
 import { SubtreeUpdater } from "@nocturne-xyz/subtree-updater";
 import RedisMemoryServer from "redis-memory-server";
@@ -109,7 +109,7 @@ describe("Wallet, Context, Bundler, and SubtreeUpdater", async () => {
       nocturneContextAlice,
       notesDBBob,
       nocturneContextBob,
-    } = await setupNocturne({ deployContracts: true }));
+    } = await setupNocturne(deployer));
 
     const serverDB = open({ path: `${__dirname}/../db/localMerkleTestDB` });
     const prover = getSubtreeUpdateProver();